feat : 생년월일 select 데이터 생성
